pyproject.toml
[build-system]
requires = ["hatchling>=1.10.0"]
build-backend = "hatchling.build"

[project]
name = "api-ninja"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "allure-pytest>=2.14.1",
    "openai-agents>=0.0.14",
    "openai>=1.76.0",
    "pytest>=8.3.5",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "pydantic>=2.11.4",
    "rich>=14.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.11.7",
]

[project.urls]
Homepage = "https://github.com/yourorg/api-ninja"

[project.entry-points."pytest11"]
api_ninja = "api_ninja.plugin"

[tool.uv.sources]
api-ninja = { path = "./src/foo" }
# api-ninja = { workspace = true }

# [tool.uv.workspace]
# members = ["src/*"]
